Fixed structure page_objects, add some test
