+++ run_tests.py
@@ -2,12 +2,12 @@
 
 import unittest
 import sys
-from tests import SimpleTest, DebugTest, AnswersTest
+from tests import DebugTest, AnswersTest
 
 if __name__ == '__main__':
     suite = unittest.TestSuite((
-        unittest.makeSuite(AnswersTest),
-        # unittest.makeSuite(DebugTest),
+        # unittest.makeSuite(AnswersTest),
+        unittest.makeSuite(DebugTest),
     ))
     result = unittest.TextTestRunner().run(suite)
     sys.exit(not result.wasSuccessful())
